Move output converters to a shared module
